fix(*): continueAs doesn't have invoke & onParentComplete field
