+++ model/workflow_test.go
@@ -15,9 +15,12 @@
 package model
 
 import (
-	val "github.com/serverlessworkflow/sdk-go/v2/validator"
-	"github.com/stretchr/testify/assert"
+	"encoding/json"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	val "github.com/serverlessworkflow/sdk-go/v2/validator"
 )
 
 func TestContinueAsStructLevelValidation(t *testing.T) {
@@ -31,13 +34,9 @@ func TestContinueAsStructLevelValidation(t *testing.T) {
 		{
 			name: "valid ContinueAs",
 			continueAs: ContinueAs{
-				WorkflowRef: WorkflowRef{
-					WorkflowID:       "another-test",
-					Version:          "2",
-					Invoke:           "sync",
-					OnParentComplete: "terminate",
-				},
-				Data: "${ del(.customerCount) }",
+				WorkflowID: "another-test",
+				Version:    "2",
+				Data:       "${ del(.customerCount) }",
 				WorkflowExecTimeout: WorkflowExecTimeout{
 					Duration:  "PT1H",
 					Interrupt: false,
@@ -49,11 +48,9 @@ func TestContinueAsStructLevelValidation(t *testing.T) {
 		{
 			name: "invalid WorkflowExecTimeout",
 			continueAs: ContinueAs{
-				WorkflowRef: WorkflowRef{
-					WorkflowID: "test",
-					Version:    "1",
-				},
-				Data: "${ del(.customerCount) }",
+				WorkflowID: "test",
+				Version:    "1",
+				Data:       "${ del(.customerCount) }",
 				WorkflowExecTimeout: WorkflowExecTimeout{
 					Duration: "invalid",
 				},
@@ -72,6 +69,82 @@ func TestContinueAsStructLevelValidation(t *testing.T) {
 				return
 			}
 			assert.NoError(t, err)
+		})
+	}
+}
+
+func TestContinueAsUnmarshalJSON(t *testing.T) {
+	type testCase struct {
+		desp   string
+		data   string
+		expect ContinueAs
+		err    string
+	}
+	testCases := []testCase{
+		{
+			desp: "string",
+			data: `"1"`,
+			expect: ContinueAs{
+				WorkflowID: "1",
+			},
+			err: ``,
+		},
+		{
+			desp: "object all field set",
+			data: `{"workflowId": "1", "version": "2", "data": "3", "workflowExecTimeout": {"duration": "PT1H", "interrupt": true, "runBefore": "4"}}`,
+			expect: ContinueAs{
+				WorkflowID: "1",
+				Version:    "2",
+				Data:       "3",
+				WorkflowExecTimeout: WorkflowExecTimeout{
+					Duration:  "PT1H",
+					Interrupt: true,
+					RunBefore: "4",
+				},
+			},
+			err: ``,
+		},
+		{
+			desp: "object optional field unset",
+			data: `{"workflowId": "1"}`,
+			expect: ContinueAs{
+				WorkflowID: "1",
+				Version:    "",
+				Data:       nil,
+				WorkflowExecTimeout: WorkflowExecTimeout{
+					Duration:  "",
+					Interrupt: false,
+					RunBefore: "",
+				},
+			},
+			err: ``,
+		},
+		{
+			desp:   "invalid string format",
+			data:   `"{`,
+			expect: ContinueAs{},
+			err:    `unexpected end of JSON input`,
+		},
+		{
+			desp:   "invalid object format",
+			data:   `{"workflowId": 1}`,
+			expect: ContinueAs{},
+			err:    `json: cannot unmarshal number into Go struct field continueAsForUnmarshal.workflowId of type string`,
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.desp, func(t *testing.T) {
+			var v ContinueAs
+			err := json.Unmarshal([]byte(tc.data), &v)
+
+			if tc.err != "" {
+				assert.Error(t, err)
+				assert.Regexp(t, tc.err, err)
+				return
+			}
+
+			assert.NoError(t, err)
+			assert.Equal(t, tc.expect, v)
 		})
 	}
 }
